Move rootReducer to /src/redux folder
